@@ -3,7 +3,7 @@ source 'https://rubygems.org'
 # Run `bundle config --global github.https true` on the server to quiet warnings
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.2.9'
+gem 'rails', '~> 5.2'
 
 # Use postgresql as the database for Active Record
 gem 'pg', '~> 0.21.0'
@@ -24,13 +24,13 @@ gem 'jquery-rails', '~> 4.3.1'
 # Use modernizr for browser feature detection
 gem 'modernizr-rails', '~> 2.7.0'
 # Use the NYU Libraries assets gem
-gem 'nyulibraries_stylesheets', github: 'NYULibraries/nyulibraries_stylesheets', tag: 'v1.1.0'
-gem 'nyulibraries_templates', github: 'NYULibraries/nyulibraries_templates', tag: 'v1.2.0'
+gem 'nyulibraries_stylesheets', github: 'NYULibraries/nyulibraries_stylesheets', tag: 'v1.1.2'
+gem 'nyulibraries_templates', github: 'NYULibraries/nyulibraries_templates', tag: 'v1.2.2'
 gem 'nyulibraries_institutions', github: 'NYULibraries/nyulibraries_institutions', tag: 'v1.0.3'
 gem 'nyulibraries_javascripts', github: 'NYULibraries/nyulibraries_javascripts', tag: 'v1.0.0'
-gem 'nyulibraries_errors', github: 'NYULibraries/nyulibraries_errors', tag: 'v1.0.1'
+gem 'nyulibraries_errors', github: 'NYULibraries/nyulibraries_errors', tag: 'v1.1.0'
 # Use higher version of Compass CSS framework for sprites, etc.
-gem 'compass-rails', '~> 3.0.0'
+gem 'compass-rails', '~> 3.1'
 
 # Use the NYU Libraries deploy gem
 gem 'formaggio', github: 'NYULibraries/formaggio', tag: 'v1.7.2'
@@ -62,6 +62,8 @@ gem 'font-awesome-rails', '~> 4'
 # NOTE: factory_bot_rails fix needed before update to >4.3.1
 # see issue: https://github.com/doorkeeper-gem/doorkeeper/issues/1043
 gem 'doorkeeper', '~> 4.2.0'
+# Manually include responders to maintain respond_with & respond_to functionality
+gem 'responders', '~> 2.0'
 
 # Figs for configuration
 gem 'figs', '~> 2.1'
@@ -72,6 +74,8 @@ gem 'faraday_middleware', '~> 0.12'
 
 gem 'dalli', '~> 2.7.8'
 
+gem 'bootsnap', require: false
+
 group :doc do
   # bundle exec rake doc:rails generates the API under doc/api.
   gem 'sdoc', require: false
@@ -81,32 +85,38 @@ end
 group :development do
   gem 'better_errors', '~> 2', platform: :ruby
   gem 'binding_of_caller', '~> 0.8', platform: :ruby
+  gem 'listen'
 end
 
 # Testing gems
 group :test, :cucumber do
   # Coveralls for testing coverage
   gem 'coveralls', '~> 0.8', require: false
-  gem 'cucumber-rails', '~> 1.5', require: false
+  gem 'cucumber-rails', '~> 1.6', require: false
   gem 'selenium-webdriver', '~> 3'
   gem 'pickle', '~> 0.5'
   gem 'database_cleaner', '~> 1.7'
   gem 'vcr', '~> 4'
   gem 'webmock', '~> 3'
   # Rspec as the test framework
-  gem 'rspec-rails', '~> 3.6'
+  gem 'rspec-rails', '~> 3.7'
   # Phantomjs for headless browser testing
   gem 'phantomjs', '>= 1.9.7'
   gem 'poltergeist', '~> 1'
   # Use factory_bot for creating models
-  gem 'factory_bot_rails', '~> 4.8.2'
+  gem 'factory_bot_rails', '~> 4.9'
   # Use json_spec to do rspec tests with JSON
   gem 'json_spec', '~> 1.1'
   gem 'rspec-its', '~> 1.2'
   gem 'faker', '~> 1'
+  # allows for assigns and assert_template testing in Rails 5
+  gem 'rails-controller-testing'
 end
 
 group :test, :cucumber, :development do
-  gem 'pry', '~> 0'
-  gem 'pry-remote', '~> 0'
+  # Use Puma as the app server for testing and local development
+  gem 'puma', '~> 3.12'
+
+  gem 'pry-rails'
+  gem 'byebug'
 end
