@@ -1,8 +1,8 @@
 module AuthTypeHelper
 
   def auth_type_with_institution
-    return "#{params["auth_type"]}.#{params["institution"]}" if ["nyush","nyuad"].include?(params["institution"])
-    return "#{params["auth_type"]}"
+    suffix = ["nyush", "nyuad"].include?(params[:institution]) ? ".#{params[:institution]}" : ""
+    "#{params[:auth_type]}#{suffix}"
   end
 
   # Return a help-block for the password field if the institution has a
