@@ -2,7 +2,7 @@
 ENV['RAILS_ENV'] ||= 'test'
 require 'coveralls'
 Coveralls.wear_merged!('rails')
-require File.expand_path('../../config/environment', __FILE__)
+require_relative '../config/environment'
 # Prevent database truncation if the environment is production
 abort("The Rails environment is running in production mode!") if Rails.env.production?
 require 'rails_helper'
