@@ -1,8 +1,8 @@
 require 'rails_helper'
 describe "NYU Shibboleth Login" do
-  before { https!; get("/users/auth/nyu_shibboleth/callback?auth_type=nyu", nil, environment) }
+  before { https!; get("/users/auth/nyu_shibboleth/callback?auth_type=nyu", { params: environment }) }
   context "when there is no Shibboleth session in the SP " do
-    let(:environment) { nil }
+    let(:environment) { {} }
     describe 'OmniAuth::AuthHash' do
       subject(:omniauth_hash) { @request.env['omniauth.auth'] if @request.present? }
       it { should be_blank }
